+++ python/GafferSceneUI/__init__.py
@@ -172,6 +172,7 @@ from . import BoundQueryUI
 from . import ExistenceQueryUI
 from . import AttributeQueryUI
 from . import UVSamplerUI
+from . import CryptomatteUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.