+++ src/GafferSceneModule/CryptomatteBinding.cpp
@@ -1,14 +1,13 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2012-2014, Image Engine Design Inc. All rights reserved.
-//  Copyright (c) 2013, John Haddon. All rights reserved.
+//  Copyright (c) 2021, Murray Stevenson. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
 //  met:
 //
 //      * Redistributions of source code must retain the above
-//        copyright notice, this list of conditions and the following
+//       copyright notice, this list of conditions and the following
 //        disclaimer.
 //
 //      * Redistributions in binary form must reproduce the above
@@ -37,58 +36,28 @@
 
 #include "boost/python.hpp"
 
-#include "AttributesBinding.h"
-#include "CoreBinding.h"
-#include "EditScopeAlgoBinding.h"
-#include "FilterBinding.h"
-#include "GlobalsBinding.h"
-#include "HierarchyBinding.h"
-#include "IECoreGLPreviewBinding.h"
-#include "IOBinding.h"
-#include "TweaksBinding.h"
-#include "ObjectProcessorBinding.h"
-#include "OptionsBinding.h"
-#include "PrimitiveSamplerBinding.h"
-#include "PrimitiveVariablesBinding.h"
-#include "PrimitivesBinding.h"
-#include "RenderBinding.h"
-#include "RenderControllerBinding.h"
-#include "SceneAlgoBinding.h"
-#include "ScenePathBinding.h"
-#include "SetAlgoBinding.h"
-#include "ShaderBinding.h"
-#include "TransformBinding.h"
-#include "QueryBinding.h"
 #include "CryptomatteBinding.h"
 
+#include "GafferScene/Cryptomatte.h"
+
+#include "GafferBindings/DependencyNodeBinding.h"
+#include "GafferBindings/PlugBinding.h"
+
 using namespace boost::python;
-using namespace GafferSceneModule;
+using namespace IECorePython;
+using namespace Gaffer;
+using namespace GafferBindings;
+using namespace GafferScene;
 
-BOOST_PYTHON_MODULE( _GafferScene )
+void GafferSceneModule::bindCryptomatte()
 {
-
-	bindCore();
-	bindFilter();
-	bindTransform();
-	bindGlobals();
-	bindOptions();
-	bindAttributes();
-	bindSceneAlgo();
-	bindSetAlgo();
-	bindPrimitives();
-	bindScenePath();
-	bindShader();
-	bindRender();
-	bindRenderController();
-	bindHierarchy();
-	bindObjectProcessor();
-	bindPrimitiveVariables();
-	bindTweaks();
-	bindIO();
-	bindPrimitiveSampler();
-	bindIECoreGLPreview();
-	bindEditScopeAlgo();
-	bindQueries();
-	bindCryptomatte();
-
+	{
+		scope s = GafferBindings::DependencyNodeClass<Cryptomatte>();
+		enum_<Cryptomatte::ManifestSource>( "ManifestSource" )
+			.value( "None", Cryptomatte::ManifestSource::None )
+			.value( "None_", Cryptomatte::ManifestSource::None )
+			.value( "Metadata", Cryptomatte::ManifestSource::Metadata )
+			.value( "Sidecar", Cryptomatte::ManifestSource::Sidecar )
+		;
+	}
 }