+++ startup/gui/nodeEditor.py
@@ -42,6 +42,7 @@ def __toolMenu( nodeEditor, node, menuDefinition ) :
 	GafferUI.UIEditor.appendNodeEditorToolMenuDefinitions( nodeEditor, node, menuDefinition )
 	GafferUI.BoxUI.appendNodeEditorToolMenuDefinitions( nodeEditor, node, menuDefinition )
 	GafferSceneUI.FilteredSceneProcessorUI.appendNodeEditorToolMenuDefinitions( nodeEditor, node, menuDefinition )
+	GafferSceneUI.CryptomatteUI.appendNodeEditorToolMenuDefinitions( nodeEditor, node, menuDefinition )
 
 GafferUI.NodeEditor.toolMenuSignal().connect( __toolMenu, scoped = False )